project_dj/assets/new_project/test_requirements.txt
-r requirements.txt
coverage
django_nose
nose
nose_exclude
nosexcover
pylint
